cleanup: cleanup tools
* merge tools/price.js into tools/sync.js
* price ticker enabled by default
* remove unused patcher.js and blockHelper.js from tools folder
