+++ config.example.json
@@ -26,7 +26,6 @@
         "poweredbyCustom": false,
         "poweredbyEtcImage": "/img/powered-by-etcexplorer-w.png",
         "poweredbyEtc": true,
-        "useFiat":      false,
         "tokenList": "tokens.json",
         "miners": {
             "0xdf7d7e053933b5cc24372f878c90e62dadad5d42": "EtherMine",
