+++ public/views/address.html
@@ -13,7 +13,7 @@
                 <span class="eth-stat-text">
                   {{ settings.symbol }} Balance <i class="fa fa-question-circle" data-toggle="tooltip" data-placement="right" title="Note: Balance may not reflect transaction data if you have transactions resulting from Contract Internal Transactions. We are working on adding that functionality."></i>
                 </span><br>
-                <div ng-show="settings.useFiat" class="margin-top-20">
+                <div class="margin-top-20">
                   <span class="fade">{{ addr.balanceUSD | currency : "$" : 4 }}</span> <span class="eth-stat-text fade">USD balance</span>
                 </div>
                 <div ng-hide="addr.mined==0" class="margin-top-20">
