--- tools/blockHelper.js
@@ -1,207 +1,0 @@
-require( '../db.js' );
-var etherUnits = require("../lib/etherUnits.js");
-var BigNumber = require('bignumber.js');
-
-var Web3 = require('web3');
-
-var mongoose        = require( 'mongoose' );
-var Block           = mongoose.model( 'Block' );
-var Transaction     = mongoose.model( 'Transaction' );
-
-var grabBlocks = function(config) {
-    if('listenOnly' in config && config.listenOnly === true)
-        listenBlocks(config, web3);
-    else
-        setTimeout(function() {
-            grabBlock(config, web3, config.blocks.pop());
-        }, 2000);
-}
-
-var listenBlocks = function(config, web3) {
-    var newBlocks = web3.eth.filter("latest");
-    newBlocks.watch(function (error, log) {
-        if(error) {
-            console.log('Error: ' + error);
-        } else if (log == null) {
-            console.log('Warning: null block hash');
-        } else {
-            grabBlock(config, web3, log);
-        }
-    });
-}
-
-var grabBlock = function(config, web3, blockHashOrNumber) {
-    var desiredBlockHashOrNumber;
-    // check if done
-    if(blockHashOrNumber == undefined) {
-        return;
-    }
-    if (typeof blockHashOrNumber === 'object') {
-        if('start' in blockHashOrNumber && 'end' in blockHashOrNumber) {
-            desiredBlockHashOrNumber = blockHashOrNumber.end;
-        }
-        else {
-            console.log('Error: Aborted becasue found a interval in blocks ' +
-                'array that doesn\'t have both a start and end.');
-            process.exit(9);
-        }
-    }
-    else {
-        desiredBlockHashOrNumber = blockHashOrNumber;
-    }
-    if(web3.eth.net.isListening()) {
-        web3.eth.getBlock(desiredBlockHashOrNumber, true, function(error, blockData) {
-            if(error) {
-                console.log('Warning: error on getting block with hash/number: ' +
-                    desiredBlockHashOrNumber + ': ' + error);
-            }
-            else if(blockData == null) {
-                console.log('Warning: null block data received from the block with hash/number: ' +
-                    desiredBlockHashOrNumber);
-            }
-            else {
-                if('terminateAtExistingDB' in config && config.terminateAtExistingDB === true) {
-                    checkBlockDBExistsThenWrite(config, blockData);
-                }
-                else {
-                    writeBlockToDB(config, blockData);
-                }
-                if (!('skipTransactions' in config && config.skipTransactions === true))
-                    writeTransactionsToDB(config, blockData);
-                if('listenOnly' in config && config.listenOnly === true)
-                    return;
-
-                if('hash' in blockData && 'number' in blockData) {
-                    // If currently working on an interval (typeof blockHashOrNumber === 'object') and
-                    // the block number or block hash just grabbed isn't equal to the start yet:
-                    // then grab the parent block number (<this block's number> - 1). Otherwise done
-                    // with this interval object (or not currently working on an interval)
-                    // -> so move onto the next thing in the blocks array.
-                    if(typeof blockHashOrNumber === 'object' &&
-                        (
-                            (typeof blockHashOrNumber['start'] === 'string' && blockData['hash'] !== blockHashOrNumber['start']) ||
-                            (typeof blockHashOrNumber['start'] === 'number' && blockData['number'] > blockHashOrNumber['start'])
-                        )
-                    ) {
-                        blockHashOrNumber['end'] = blockData['number'] - 1;
-                        grabBlock(config, web3, blockHashOrNumber);
-                    }
-                    else {
-                        grabBlock(config, web3, config.blocks.pop());
-                    }
-                }
-                else {
-                    console.log('Error: No hash or number was found for block: ' + blockHashOrNumber);
-                    process.exit(9);
-                }
-            }
-        });
-    }
-    else {
-        console.log('Error: Aborted due to web3 is not connected when trying to ' +
-            'get block ' + desiredBlockHashOrNumber);
-        process.exit(9);
-    }
-}
-
-var writeBlockToDB = function(config, blockData) {
-    return new Block(blockData).save( function( err, block, count ){
-        if ( typeof err !== 'undefined' && err ) {
-            if (err.code == 11000) {
-                console.log('Skip: Duplicate key ' +
-                blockData.number.toString() + ': ' +
-                err);
-            } else {
-               console.log('Error: Aborted due to error on ' +
-                    'block number ' + blockData.number.toString() + ': ' +
-                    err);
-               process.exit(9);
-           }
-        } else {
-            if(!('quiet' in config && config.quiet === true)) {
-                console.log('DB successfully written for block number ' +
-                    blockData.number.toString() );
-            }
-        }
-      });
-}
-
-/**
-  * Checks if the a record exists for the block number then ->
-  *     if record exists: abort
-  *     if record DNE: write a file for the block
-  */
-var checkBlockDBExistsThenWrite = function(config, blockData) {
-    Block.find({number: blockData.number}, function (err, b) {
-        if (!b.length)
-            writeBlockToDB(config, blockData);
-        else {
-            console.log('Aborting because block number: ' + blockData.number.toString() +
-                ' already exists in DB.');
-            process.exit(9);
-        }
-
-    })
-}
-
-/**
-    Break transactions out of blocks and write to DB
-**/
-var writeTransactionsToDB = function(config, blockData) {
-    var bulkOps = [];
-    if (blockData.transactions.length > 0) {
-        for (d in blockData.transactions) {
-            var txData = blockData.transactions[d];
-            txData.timestamp = blockData.timestamp;
-            txData.value = etherUnits.toEther(new BigNumber(txData.value), 'wei');
-            bulkOps.push(txData);
-        }
-        Transaction.collection.insert(bulkOps, function( err, tx ){
-            if ( typeof err !== 'undefined' && err ) {
-                if (err.code == 11000) {
-                    console.log('Skip: Duplicate key ' +
-                    err);
-                } else {
-                   console.log('Error: Aborted due to error: ' +
-                        err);
-                   process.exit(9);
-               }
-            } else if(!('quiet' in config && config.quiet === true)) {
-                console.log('DB successfully written for block ' +
-                    blockData.transactions.length.toString() );
-            }
-        });
-    }
-}
-
-// load config.json
-var config = { nodeAddr: 'localhost', wsPort: 8546, bulkSize: 100 };
-try {
-    var local = require('../config.json');
-    _.extend(config, local);
-    console.log('config.json found.');
-} catch (error) {
-    if (error.code === 'MODULE_NOT_FOUND') {
-        var local = require('../config.example.json');
-        _.extend(config, local);
-        console.log('No config file found. Using default configuration... (config.example.json)');
-    } else {
-        throw error;
-        process.exit(1);
-    }
-}
-
-console.log('Connecting ' + config.nodeAddr + ':' + config.wsPort + '...');
-
-var web3 = new Web3(new Web3.providers.WebsocketProvider('ws://' + config.nodeAddr + ':' + config.wsPort.toString()));
-
-// set the default blocks if it's not provided
-if (!('blocks' in config) || !(Array.isArray(config.blocks))) {
-    config.blocks = [];
-    config.blocks.push({'start': 0, 'end': 'latest'});
-}
-
-console.log('Using configuration:');
-console.log(config);
-
-grabBlocks(config);
